edited 44

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,23 +5,23 @@
  */
 int main(void)
 {
-uint64_t i = 1;
-uint64_t j = 2;
+long double i = 1;
+long double j = 2;
 int f = 0;
-uint64_t h, k1, k2;
+long double h, k1, k2;
 
 	while (f < 98)
 	{
 		if (f == 0)
 		{
-			printf("%lu", i);
+			printf("%Lf", i);
 			printf(", ");
 		}
 		else if (f == 1)
 		{
 			k1 = i;
 			k2 = j;
-			printf("%lu", j);
+			printf("%Lf", j);
 			printf(", ");
 		}
 		else
@@ -29,7 +29,7 @@ uint64_t h, k1, k2;
 			h = k1 + k2;
 			k1 = k2;
 			k2 = h;
-			printf("%lu", h);
+			printf("%Lf", h);
 
 			if (f != 97)
 			{
